requirements.txt
fastapi>=0.115.0
sqlalchemy>=2.0.36
asyncpg>=0.30.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.17
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
email-validator>=2.2.0
resend>=2.0.0
mangum>=0.17.0
bcrypt>=4.0.0
